sdk/python/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "runner-sdk"
version = "1.0.0"
description = "Python SDK for Runner System"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "httpx>=0.25.0",
    "sseclient-py>=1.8.0",
]

[project.optional-dependencies]
dev = [
    "openapi-python-client>=0.16.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["runner_sdk*"]
